Refactor and bug fixes
