+++ src/components/discgolfscorecard/Hole.java
@@ -1,3 +1,5 @@
+package components.discgolfscorecard;
+
 public record Hole(int par, int distance, int strokes) {
 
 }
