+++ test/components/discgolfscorecard/DiscGolfScorecardKernelTest.java
@@ -1,3 +1,5 @@
+package components.discgolfscorecard;
+
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
