+++ test/components/discgolfscorecard/DiscGolfScorecardTest.java
@@ -1,3 +1,5 @@
+package components.discgolfscorecard;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
